Added configuration changes
configurations for node js deployment

production changes to git ignore

update api documentation

added files for front end and database

removed uneeded comments

folder change for nodeJS api
