+++ .gitignore
@@ -1,3 +1,2 @@
-**/nba-api/node_modules
-**/nba-api/package-lock.json
-**/nba-api/package.json
+**/nba-api-nodeJS/node_modules
+**/nba-api-nodeJS/package-lock.json
