+++ nba-api-nodeJS/.env
@@ -8,3 +8,4 @@ AZURE_DB_USER="nbaAdmin"
 AZURE_DB_PASSWORD="nba_prediction1"
 AZURE_DB_DATABASE="NBA"
 AZURE_DIALECT="mssql"
+PORT=3070
