+++ nba-api-nodeJS/controller-functions/player-controller-functions.js
@@ -1,25 +1,28 @@
-const { text } = require("express");
-const { sequelize } = require("../db-config");
-
-
-const db = require('../db-config')
+const {connectToDB, createDBContext} = require("../db-config");
 
 //for later use useful for searching
 //const Op = db.Sequelize.Op;
 
-
+const Sequelize = require("sequelize");
 
 exports.getPlayers = async (req, res) => {
-  const players = await  db.Players.findAll({
+
+  const sequelize = connectToDB();
+
+  const db = createDBContext(sequelize);
+
+  db.Players = require("../models/players.js")(sequelize, Sequelize);
+
+  await  db.Players.findAll({
     attributes: ['PlayerID','FirstName','LastName','Year','Wins','Losses','PlayerWinPercentage','Points','Rebounds','Assists','Steals','Blocks','MissedFieldGoals','MissedFreeThrows','TurnOvers']
   })
     .then(
-      data => { res.send(data), console.log(data)}
+      data => { res.send(data)}
     )
     .catch(
       err => res.status(500)
         .send({ message: err.message || "internal server error" })
   );
-  
-  db.close();
+
+  sequelize.close();
 }
